NavigationController with Segue
The screens presented by a navigation interface typically mimic the
hierarchical organization of your data. At each level of the hierarchy,
you provide an appropriate screen (managed by a custom view controller)
to display the content at that level. Figure 1 shows an example of the
navigation interface presented by the Settings application in iOS
Simulator. The first screen presents the user with the list of
applications that contain preferences. Selecting an application reveals
individual settings and groups of settings for that application.
Selecting a group yields more settings and so on. For all but the root
view, the navigation controller provides a back button to allow the
user to move back up the hierarchy.